+++ org/diceresearch/nebula/translation/translator.py
@@ -1,5 +1,5 @@
 import threading
-import httpmanager
+from database import httpmanager
 import settings
 import orchestrator
